Remove blocking from InMemory
