+++ packages/in-memory/tests/go.spec.ts
@@ -130,54 +130,4 @@ describe("go", () => {
       });
     });
   });
-
-  describe("with user confirmation", () => {
-    it("calls response handler after the user confirms the navigation", done => {
-      const testHistory = InMemory();
-      let setup = false;
-      const router = ignoreFirstCall(function(pending) {
-        pending.finish();
-        if (!setup) {
-          return;
-        }
-        expect(testHistory.location).toMatchObject({
-          pathname: "/",
-          key: "0.0"
-        });
-        done();
-      });
-      const confirm = (info, confirm, prevent) => {
-        confirm();
-      };
-      testHistory.confirmWith(confirm);
-      testHistory.respondWith(router);
-
-      testHistory.navigate("/two", "push"); // 1.0
-      testHistory.navigate("/three", "push"); // 2.0
-      setup = true;
-      testHistory.go(-2);
-    });
-
-    it("does not call response handler when the user prevents the navigation", () => {
-      const testHistory = InMemory();
-      const confirm = (info, confirm, prevent) => {
-        prevent();
-      };
-      function router(pending) {
-        pending.finish();
-      }
-      testHistory.respondWith(router);
-
-      testHistory.navigate("/two", "push"); // 1.0
-      testHistory.navigate("/three", "push"); // 2.0
-      // don't add function until we have setup the history
-      testHistory.confirmWith(confirm);
-      testHistory.go(-2);
-
-      expect(testHistory.location).toMatchObject({
-        pathname: "/three",
-        key: "2.0"
-      });
-    });
-  });
 });
