+++ packages/in-memory/types/InMemory.d.ts
@@ -1,0 +1,2 @@
+import { Options, InMemoryHistory } from "./types";
+export declare function InMemory<Q = string>(options?: Options<Q>): InMemoryHistory<Q>;
